Move some teams stuff around.
